requirements.txt
streamlit==1.37.1
pandas
google-cloud-translate==3.16.0
openai==1.37.0
python-dotenv==0.21.0